+++ OpenMarket/OpenMarket/Network/URLRequestProtocol.swift
@@ -7,15 +7,9 @@
 
 import Foundation
 
-protocol URLRequestProtocol: MultiPartProtocol {
+protocol URLRequestProtocol {
   static var boundary: String { get }
   func makeURLRequest(httpMethod: HttpMethod, apiRequestType: RequestType) -> URLRequest?
-  func setMultiPartBody(httpMethod: HttpMethod,
-                        apiRequestType: RequestType,
-                        product: Uploadable) -> URLRequest?
-  func setJsonBody(httpMethod: HttpMethod,
-                   apiRequestType: RequestType,
-                   product: ProductDeleteRequest) -> URLRequest?
 }
 
 extension URLRequestProtocol {
@@ -31,36 +25,6 @@ extension URLRequestProtocol {
     urlRequest.httpMethod = httpMethod.description
     urlRequest.setValue(contentType, forHTTPHeaderField: HttpMethod.contentType)
     
-    return urlRequest
-  }
-  
-  func setMultiPartBody(httpMethod: HttpMethod,
-                        apiRequestType: RequestType,
-                        product: Uploadable) -> URLRequest? {
-    guard var urlRequest = makeURLRequest(httpMethod: httpMethod,
-                                          apiRequestType: apiRequestType) else { return nil }
-    var data: Data = Data()
-    
-    product.parameters.forEach { key, value in
-      if let images = value as? [Data] {
-        data.append(makeBodyForImage(boundary: Self.boundary, parameter: key, images: images))
-      } else if let value = value {
-        data.append(makeBodyForNormal(boundary: Self.boundary, parameter: key, value: value))
-      }
-    }
-    
-    data.appendString("--\(Self.boundary)--\r\n")
-    urlRequest.httpBody = data
-    return urlRequest
-  }
-  
-  func setJsonBody(httpMethod: HttpMethod, apiRequestType: RequestType,
-                   product: ProductDeleteRequest) -> URLRequest? {
-    guard var urlRequest = makeURLRequest(httpMethod: httpMethod,
-                                          apiRequestType: apiRequestType) else { return nil }
-    guard let encodedData = try? JSONEncoder().encode(product) else { return nil }
-    
-    urlRequest.httpBody = encodedData
     return urlRequest
   }
 }
